#4712 - renaming files etc
