+++ sirepo/feature_config.py
@@ -23,7 +23,7 @@ PROD_FOSS_CODES = frozenset(
         "genesis",
         "jspec",
         "madx",
-        "ml",
+        "activait",
         "opal",
         "radia",
         "shadow",
