+++ sirepo/package_data/static/html/activait-analysis.html
