+++ sirepo/package_data/static/html/activait-classification.html
