+++ sirepo/package_data/static/html/activait-data.html
