+++ sirepo/package_data/static/html/activait-partition.html
