+++ sirepo/package_data/static/json/schema-common.json
@@ -4,6 +4,10 @@
             "longName": "MyApp",
             "shortName": "MyApp"
         },
+        "activait": {
+            "longName": "Activait",
+            "shortName": "Activait"
+        },
         "controls": {
             "longName": "Controls",
             "shortName": "Controls"
@@ -39,10 +43,6 @@
         "madx": {
             "longName": "MAD-X",
             "shortName": "MAD-X"
-        },
-        "ml": {
-            "longName": "Activait",
-            "shortName": "Activait"
         },
         "opal": {
             "longName": "OPAL",
