+++ sirepo/package_data/template/activait/classification-base.py.jinja
