+++ sirepo/package_data/template/activait/examples/happiness.json
@@ -248,5 +248,5 @@
             "notes": ""
         }
     },
-    "simulationType": "ml"
+    "simulationType": "activait"
 }
