+++ sirepo/package_data/template/activait/examples/heartDisease.json
@@ -316,5 +316,5 @@
             "notes": ""
         }
     },
-    "simulationType": "ml"
+    "simulationType": "activait"
 }
