+++ sirepo/package_data/template/activait/examples/iris.json
@@ -152,5 +152,5 @@
             "notes": ""
         }
     },
-    "simulationType": "ml"
+    "simulationType": "activait"
 }
