+++ sirepo/package_data/template/activait/knn.py.jinja
