+++ sirepo/package_data/template/activait/linear-svc.py.jinja
