+++ sirepo/package_data/template/activait/logistic-regression.py.jinja
