+++ sirepo/package_data/template/activait/partition.py.jinja
