+++ sirepo/package_data/template/activait/scale.py.jinja
