+++ sirepo/package_data/template/activait/train.py.jinja
